move chordgen function to cli.py
